--- storybook-static/4.ddeff097.iframe.bundle.js.LICENSE.txt
@@ -1,8 +1,0 @@
-/**
- * Prism: Lightweight, robust, elegant syntax highlighting
- *
- * @license MIT <https://opensource.org/licenses/MIT>
- * @author Lea Verou <https://lea.verou.me>
- * @namespace
- * @public
- */
