+++ storybook-static/iframe.html
@@ -131,4 +131,4 @@
             
         
             
-              window['FRAMEWORK_OPTIONS'] = {};</script><script src="runtime~main.cafe249d.iframe.bundle.js"></script><script src="vendors~main.805927c8.iframe.bundle.js"></script><script src="main.55ebf835.iframe.bundle.js"></script></body></html>
+              window['FRAMEWORK_OPTIONS'] = {};</script><script src="runtime~main.07dd4f5e.iframe.bundle.js"></script><script src="vendors~main.4e9d7224.iframe.bundle.js"></script><script src="main.4067c13c.iframe.bundle.js"></script></body></html>
