+++ tsconfig.json
@@ -1,54 +1,53 @@
 {
-    "compilerOptions": {
-      "target": "ES2018",
-      "outDir": "build/esm",
-      "rootDir": "src",
-      "moduleResolution": "node",
-      "module": "ES6",
-      "forceConsistentCasingInFileNames": true,
-      "inlineSourceMap": false,
-  
-      // Ensure that .d.ts files are created by tsc, but not .js files
-      "declaration": true,
-  
-      // Ensure that Babel can safely transpile files in the TypeScript project
-      "isolatedModules": true,
-  
-      /* Enables emit interoperability between CommonJS and ES Modules via creation of
+  "compilerOptions": {
+    "target": "ES2018",
+    "outDir": "build/esm",
+    "rootDir": "src",
+    "moduleResolution": "node",
+    "module": "ES6",
+    "forceConsistentCasingInFileNames": true,
+    "inlineSourceMap": false,
+
+    // Ensure that .d.ts files are created by tsc, but not .js files
+    "declaration": true,
+
+    // Ensure that Babel can safely transpile files in the TypeScript project
+    "isolatedModules": true,
+
+    /* Enables emit interoperability between CommonJS and ES Modules via creation of
       namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
-      "esModuleInterop": true,
-  
-      /* Include modules imported with .json extension. */
-      "resolveJsonModule": true,
-  
-      /* Enable all strict type-checking options. */
-      "strict": true,
-  
-      /* Additional Checks */
-      "noUnusedLocals": true /* Report errors on unused locals. */,
-      "noUnusedParameters": true /* Report errors on unused parameters. */,
-      "noImplicitReturns": true /* Report error when not all code paths in function return a value. */,
-      "noFallthroughCasesInSwitch": true /* Report errors for fallthrough cases in switch statement. */,
-      "experimentalDecorators": true /* Adds support for lit decorators. */,
-  
-      /* Debugging Options */
-      "traceResolution": false /* Report module resolution log messages. */,
-      "listEmittedFiles": false /* Print names of generated files part of the compilation. */,
-      "listFiles": false /* Print names of files part of the compilation. */,
-      "pretty": true /* Stylize errors and messages using color and context. */,
-  
-      "lib": ["ES2016", "dom"],
-      "types": ["jest"],
-      "typeRoots": ["node_modules/@types"],
-      "plugins": [
-        {
-          "name": "ts-lit-plugin",
-          "strict": true
-        }
-      ]
-    },
-    "include": ["src/**/*.ts"],
-    "exclude": ["node_modules/**", "src/stories/**", "src/**/*.stories.ts", "src/**/*.test.ts"],
-    "compileOnSave": false
-  }
-  
+    "esModuleInterop": true,
+
+    /* Include modules imported with .json extension. */
+    "resolveJsonModule": true,
+
+    /* Enable all strict type-checking options. */
+    "strict": true,
+
+    /* Additional Checks */
+    "noUnusedLocals": true /* Report errors on unused locals. */,
+    "noUnusedParameters": true /* Report errors on unused parameters. */,
+    "noImplicitReturns": true /* Report error when not all code paths in function return a value. */,
+    "noFallthroughCasesInSwitch": true /* Report errors for fallthrough cases in switch statement. */,
+    "experimentalDecorators": true /* Adds support for lit decorators. */,
+
+    /* Debugging Options */
+    "traceResolution": false /* Report module resolution log messages. */,
+    "listEmittedFiles": false /* Print names of generated files part of the compilation. */,
+    "listFiles": false /* Print names of files part of the compilation. */,
+    "pretty": true /* Stylize errors and messages using color and context. */,
+
+    "lib": ["ES2016", "dom"],
+    "types": ["jest"],
+    "typeRoots": ["node_modules/@types"],
+    "plugins": [
+      {
+        "name": "ts-lit-plugin",
+        "strict": true
+      }
+    ]
+  },
+  "include": ["src/**/*.ts"],
+  "exclude": ["node_modules/**", "src/stories/**", "src/**/*.stories.ts", "src/**/*.test.ts"],
+  "compileOnSave": false
+}
